Store commit information in Jenkins instead of calling Bitbucket each time
Instead of using an unique key for a job, we use an unique key by build.
This is required to have a build history in bitbucket, instead of
replacing the last build, we append the new build in build information.
Also, we use BuildBadge instead of using cause description. This is a
bug when the markup formatter was plain text instead of html.