Use the V2 Bitbucket pull request comment API.
The V1 pull request comment API is deprecated. Replace with a call
to the V2 API. The V2 API is paginated and may need to be called
multiple times to get the entire set of comments.

While fixing this I noticed that while fetching pull requests the
plugin is using the v2 API there, which is also paginated. So not
all open pull requests were being retrieved.