Bugfix multiple jenkins jobs from one repo causes valid jobs to be killed as outdated jobs
This change is intended to fix a bug that occurs when using Cancel outdated jobs with more that one jenkins job referencing the same repository.  When checking for outdated jobs we are comparing the repo name and pull request but no considering the jb name.  This can cause valid jobs to be killed. I have added conditionals to compare the name of the builder ot the name of the queued or running job to avoid the problem